rename: Role 클래스 enums 패키지 안으로 이동
